fix(db): percentile_cont needs float ordering; migrations bootstrap fresh DBs; harness fixes
First live run of the db-contract harness caught four more CockroachDB
dialect failures and one structural gap:

- PERCENTILE_CONT(...) WITHIN GROUP (ORDER BY duration_ms) throws
  "unknown signature: percentile_cont_impl(decimal, int)" — the ordering
  column must be float. trend, failure-registry, and stable-corridors
  were returning degraded 503s live; baselines had the same latent bug.
  The aggregation jobs already cast (::FLOAT); the lenses now do too.

- The migration chain could not bootstrap a FRESH database: pg sends a
  multi-statement file as one implicit transaction, and CRDB rejects
  creating a partial index on a column added in the same transaction
  (000003). run-migration.mjs now splits files on top-level semicolons
  and autocommits each statement, matching golang-migrate's behavior.
  Production never hit this because it evolved incrementally — CI
  bootstraps from zero every run, so it stays fixed.

- Harness corrections: /api/v1/health path, chain_id is top-level on
  receipts (not inside interaction), ON CONFLICT for the catalog seed
  (CRDB UPSERT keys on the PK, not secondary uniques), per-run agent
  name suffixes (agent names are globally unique; a collision is a 500).

Full local run: 26 migrations from zero, 10/10 crons, 11/11 lens routes
200 + not degraded, exact count coherence, notification E2E delivered.

+++ packages/ingestion-api/src/routes/baselines.ts
@@ -101,8 +101,8 @@ app.get('/baselines/cohort', async (c) => {
               MIN(target_system_type) AS "target_system_type",
               COUNT(DISTINCT emitter_agent_id)::int AS "cohort_size",
               COUNT(*)::int AS "total_interactions",
-              PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY duration_ms)::float AS "median_duration_ms",
-              PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY duration_ms)::float AS "p95_duration_ms",
+              PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY duration_ms::FLOAT)::float AS "median_duration_ms",
+              PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY duration_ms::FLOAT)::float AS "p95_duration_ms",
               COALESCE(
                 COUNT(*) FILTER (WHERE status = 'failure')::float8 /
                 NULLIF(COUNT(*), 0)::float8, 0.0
